создан модуль для сообщений и перераспределены ф-ции по модулям
