+++ js/main.js
@@ -1,14 +1,15 @@
 // https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random
 // https://learn.javascript.ru/number
-import {setMinSum, timeSync, guestLimit} from './card-processing.js';
+import {setMinSum, timeSync, guestLimit, setSendButton, setResetButton} from './card-processing.js';
 import {disableForm} from './lock.js'
 import {createMap} from './map.js'
-import {sendForm, receiveMarkers} from './server-communication.js'
+import {receiveMarkers} from './server-communication.js';
 
 setMinSum();
 timeSync();
 guestLimit();
+setSendButton();
+setResetButton();
 disableForm();
 createMap();
 receiveMarkers();
-sendForm();
